pyproject.toml
[tool.poetry]
name = "htr-torch"
version = "0.1.0"
description = ""
authors = ["wilfred11 <wilfredvancasteren@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.2.1"
numpy = "^1.26.4"
pandas = "^2.2.1"
scikit-image = "^0.22.0"
matplotlib = "^3.8.3"
levenshtein = "^0.25.0"
opencv-python = "^4.9.0.80"
pyyaml = "^6.0.1"
torchvision = "^0.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
